requirements.txt
streamlit==1.39.0
langchain-huggingface
sentence-transformers
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.3.15
langchain-community==0.3.15
pinecone==7.3.0
langchain-pinecone==0.2.11
langchain-google-genai==2.1.10
